refactor(ddi): temporary keep xslt implementation to ensure non-regression
